✨ feat: 포스트 생성, 수정, 삭제 메서드에 사용자 ID 추가 및 권한 검증 로직 구현 (#27)
* 포스트 수정 및 삭제 시 사용자 권한 검증 로직 추가
* ErrorCode에 접근 불가능한 포스트에 대한 예외 추가

+++ src/main/java/troublog/backend/domain/trouble/service/command/PostCommandService.java
@@ -25,22 +25,24 @@ public class PostCommandService {
 	private final PostQueryService postQueryService;
 	private final PostFactory postFactory;
 
-	public PostResDto createPost(PostCreateReqDto reqDto, long userId) {
+	public PostResDto createPost(Long userId, PostCreateReqDto reqDto) {
 		Post newPost = postFactory.createPostWithRequireRelations(reqDto, userId);
 		Post savedPost = postRepository.save(newPost);
 		postFactory.establishSecondaryRelations(savedPost, reqDto);
 		return PostConverter.toResponse(savedPost);
 	}
 
-	public PostResDto updatePost(long postId, PostUpdateReqDto reqDto) {
+	public PostResDto updatePost(Long userId, Long postId, PostUpdateReqDto reqDto) {
 		Post foundPost = postQueryService.findPostById(postId);
+		PostFactory.validateAuthorized(userId, foundPost);
 		postFactory.updateRelationsIfChanged(reqDto, foundPost);
 		Post savedPost = postRepository.save(foundPost);
 		return PostConverter.toResponse(savedPost);
 	}
 
-	public void deletePost(long postId) {
+	public void deletePost(Long userId, Long postId) {
 		Post foundPost = postQueryService.findPostById(postId);
+		PostFactory.validateAuthorized(userId, foundPost);
 		postRepository.delete(foundPost);
 	}
 }
